refactor: 경로 관련 기능 Path 클래스로 역할 분리
