+++ src/main/java/wooteco/subway/service/SectionService.java
@@ -3,16 +3,12 @@ package wooteco.subway.service;
 import static wooteco.subway.domain.Sections.NEED_MERGE_SIZE;
 
 import java.util.List;
-import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Sections;
 import wooteco.subway.domain.Station;
-import wooteco.subway.dto.PathRequest;
-import wooteco.subway.dto.PathResponse;
 import wooteco.subway.dto.SectionRequest;
-import wooteco.subway.dto.StationResponse;
 import wooteco.subway.repository.SectionRepository;
 import wooteco.subway.repository.StationRepository;
 
@@ -52,23 +48,5 @@ public class SectionService {
             Section mergedSection = deletedSections.get(0).merge(deletedSections.get(1));
             sectionRepository.save(mergedSection);
         }
-    }
-
-    @Transactional(readOnly = true)
-    public PathResponse calculateMinDistance(final PathRequest pathRequest) {
-        Sections sections = new Sections(sectionRepository.findAll());
-        Station startStation = stationRepository.findById(pathRequest.getSource());
-        Station endStation = stationRepository.findById(pathRequest.getTarget());
-        List<StationResponse> stationResponses =
-                toStationResponses(sections.findShortestStations(startStation, endStation));
-        int distance = sections.calculateMinDistance(startStation, endStation);
-        int fare = sections.calculateFare(distance);
-        return new PathResponse(stationResponses, distance, fare);
-    }
-
-    private List<StationResponse> toStationResponses(final List<Station> stations) {
-        return stations.stream()
-                .map(StationResponse::new)
-                .collect(Collectors.toList());
     }
 }
