+++ src/test/java/wooteco/subway/service/SectionServiceTest.java
@@ -3,7 +3,6 @@ package wooteco.subway.service;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.Assertions.tuple;
-import static org.junit.jupiter.api.Assertions.assertAll;
 import static wooteco.subway.TestFixtures.LINE_COLOR;
 import static wooteco.subway.TestFixtures.LINE_SIX;
 import static wooteco.subway.TestFixtures.STANDARD_DISTANCE;
@@ -11,7 +10,6 @@ import static wooteco.subway.TestFixtures.동묘앞역;
 import static wooteco.subway.TestFixtures.신당역;
 import static wooteco.subway.TestFixtures.창신역;
 
-import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,15 +19,13 @@ import wooteco.subway.domain.Line;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.LineResponse;
-import wooteco.subway.dto.PathRequest;
-import wooteco.subway.dto.PathResponse;
 import wooteco.subway.dto.SectionRequest;
 import wooteco.subway.dto.StationResponse;
+import wooteco.subway.exception.SectionCreateException;
+import wooteco.subway.exception.SectionDeleteException;
 import wooteco.subway.repository.LineRepository;
 import wooteco.subway.repository.SectionRepository;
 import wooteco.subway.repository.StationRepository;
-import wooteco.subway.exception.SectionCreateException;
-import wooteco.subway.exception.SectionDeleteException;
 
 @Transactional
 @SpringBootTest
@@ -177,32 +173,6 @@ class SectionServiceTest {
 
         assertThatThrownBy(() -> sectionService.delete(lineSix.getId(), saved_동묘앞역.getId()))
                 .isInstanceOf(SectionDeleteException.class);
-
-    }
-
-    @DisplayName("구간 사이의 최소거리, 역, 비용을 반환한다")
-    @Test
-    void calculateMinDistance() {
-        Station saved_신당역 = createStation(신당역);
-        Station saved_동묘앞역 = createStation(동묘앞역);
-        Station saved_창신역 = createStation(창신역);
-
-        Line line = createLine(LINE_SIX, LINE_COLOR);
-
-        sectionRepository.save(new Section(line.getId(), saved_신당역, saved_동묘앞역, STANDARD_DISTANCE));
-        sectionRepository.save(new Section(line.getId(), saved_동묘앞역, saved_창신역, STANDARD_DISTANCE));
-
-        PathResponse pathResponse = sectionService.calculateMinDistance(new PathRequest(saved_신당역.getId(), saved_창신역.getId(), 10));
-        List<StationResponse> stations = pathResponse.getStations();
-
-        assertAll(() -> {
-            assertThat(stations).hasSize(3);
-            assertThat(stations).containsExactly(new StationResponse(saved_신당역),
-                    new StationResponse(saved_동묘앞역),
-                    new StationResponse(saved_창신역));
-            assertThat(pathResponse.getDistance()).isEqualTo(STANDARD_DISTANCE + STANDARD_DISTANCE);
-            assertThat(pathResponse.getFare()).isEqualTo(1450);
-        });
 
     }
